build(firmware): clippy every configuration the firmware compiles in
The firmware lint gate ran clippy once, `--lib` on the host. That skipped
the crate's largest tree: `adapters/`, `runtime.rs`, and `main.rs` compile
only under `cfg(target_os = "espidf")`, where nothing but the release build
saw them and that build does not deny warnings. Feature-gated code was
unchecked for the same reason, which is why an unused `serde_json::json`
import sat in `api/responses.rs` reported by every `--features api-spec`
run and caught by no gate.

`lint` now runs three clippy passes with `-D warnings`: the host crate core,
the host build with the two host-only features and their bins, and the device
target, which covers the adapters, the runtime, and the binary. The device
pass shares the release profile and target triple with `build`, so it reuses
that target directory rather than populating a second one.

Closing the blind spot exposed four warnings in the adapter tree. Three are
fixed: a redundant cast against `ESP_OK` and two ESP-IDF configuration
structs built by field reassignment after `Default::default()`. The fourth
stands: the reboot handler keeps its wrapping closure under a narrow allow,
because the bare function item cannot supply the higher-ranked connection
lifetime the handler signature requires.

+++ firmware/streamline/src/adapters/logs.rs
@@ -89,7 +89,7 @@ pub fn install() {
         // before this boot can write over them.
         unsafe { esp_log_set_vprintf(Some(capture_line)) };
         if unsafe { heap_caps_register_failed_alloc_callback(Some(note_allocation_failure)) }
-            != ESP_OK as i32
+            != ESP_OK
         {
             log::warn!("allocation failures will not be captured in the device log");
         }
